+++ mockingbird-core/src/main/java/org/qubership/automation/itf/executor/service/TCContextService.java
@@ -115,7 +115,7 @@ public class TCContextService {
         if (StringUtils.isNotEmpty(dumpfilePath)) {
             tcContext.getReportLinks().put("Download TCPDump", dumpfilePath);
         }
-        localRunningContexts.put((BigInteger) tcContext.getID(), tcContext);
+        localRunningContexts.put(tcContext.getID(), tcContext);
         CacheServices.getTcContextCacheService().set(tcContext, true);
         long stTime = System.currentTimeMillis();
         eventBusProvider.post(new TcContextEvent.Start(tcContext));
@@ -132,7 +132,7 @@ public class TCContextService {
     }
 
     public void stop(TcContext tcContext) {
-        stop((BigInteger) tcContext.getID(), getTenantId(tcContext));
+        stop(tcContext.getID(), getTenantId(tcContext));
     }
 
     public void stop(BigInteger tcContextId, String tenantId) {
@@ -147,7 +147,7 @@ public class TCContextService {
             tcContext.setEndTime(new Date());
             tcContext.setStatus(Status.STOPPED);
             performExtraFinishActions(tcContext);
-            localRunningContexts.invalidate((BigInteger) tcContext.getID());
+            localRunningContexts.invalidate(tcContext.getID());
             eventBusProvider.post(new TcContextEvent.Stop(tcContext));
         }
     }
@@ -168,7 +168,7 @@ public class TCContextService {
             log.error("Error occurred while resuming the {} context: ", tcContext.getName(), exception);
             throw exception;
         }
-        localRunningContexts.put((BigInteger) tcContext.getID(), tcContext);
+        localRunningContexts.put(tcContext.getID(), tcContext);
         CacheServices.getTcBindingCacheService().bind(tcContext);
         updateInfo(tcContext);
         ExecutionServices.getExecutionProcessManagerService().resume(tcContext);
@@ -178,7 +178,7 @@ public class TCContextService {
         Status tcContextStatus = tcContext.getStatus();
         if (Status.FAILED.equals(tcContextStatus) || Status.STOPPED.equals(tcContextStatus)) {
             tcContext.setStatus(Status.IN_PROGRESS);
-            localRunningContexts.put((BigInteger) tcContext.getID(), tcContext);
+            localRunningContexts.put(tcContext.getID(), tcContext);
             eventBusProvider.post(new TcContextEvent.Continue(tcContext));
         }
     }
@@ -187,7 +187,7 @@ public class TCContextService {
         if (Status.IN_PROGRESS.equals(tcContext.getStatus()) || Status.PAUSED.equals(tcContext.getStatus())) {
             finalizeContext(tcContext, (tcContext.isValidationFailed()) ? Status.FAILED : Status.PASSED);
             performExtraFinishActions(tcContext);
-            localRunningContexts.invalidate((BigInteger) tcContext.getID());
+            localRunningContexts.invalidate(tcContext.getID());
         }
         if (!tcContext.isFinishEventSent()) {
             synchronized (tcContext) {
@@ -211,7 +211,7 @@ public class TCContextService {
                     ? Status.FAILED_BY_TIMEOUT
                     : Status.FAILED);
             performExtraFinishActions(tcContext);
-            localRunningContexts.invalidate((BigInteger) tcContext.getID());
+            localRunningContexts.invalidate(tcContext.getID());
         }
         if (!tcContext.isFailEventSent()) {
             synchronized (tcContext) {
@@ -279,7 +279,7 @@ public class TCContextService {
 
     public void pause(TcContext tcContext) {
         executorToMessageBrokerSender.sendMessageToTcContextOperationsTopic(
-                new TcContextOperationMessage(Status.PAUSED.name(), (BigInteger) tcContext.getID()),
+                new TcContextOperationMessage(Status.PAUSED.name(), tcContext.getID()),
                 getTenantId(tcContext));
     }
 
@@ -419,7 +419,7 @@ public class TCContextService {
     @Nonnull
     private TcContext createInMemory(Object id, BigInteger projectId, UUID projectUuid) {
         TcContext context = new TcContext();
-        context.setID(id);
+        context.setID((BigInteger)id);
         context.setProjectId(projectId);
         context.setProjectUuid(projectUuid);
         context.setPodName(Config.getConfig().getRunningHostname());
